feat festivalClientContent
